pyproject.toml
[tool.poetry]
name = "ovhapi"
version = "0.1.0"
description = ""
authors = ["human dev <dev@dev.com>"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.70.1"
openstacksdk = "^0.61.0"
httpx = "^0.21.1"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
flake8 = "^4.0.1"
black = "^21.12b0"
requests = "^2.26.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"